refactor: create hero component
